Support both OSS and W3C dialects in Ruby bindings
Ruby bindings need to detect what protocol dialect it should talk
using protocol handshake which:

1. Prepares capabilities for both OSS and W3C dialects.
2. Sends them in a single payload like this:

{
  "desiredCapabilities": {...},
  "capabilities": {
    "firstMatch": [{...}]
  }
}

3. Based on a response, understand which dialect it needs to use.

This commit adds Remote::Bridge.handshake, refactors driver-specific
bridges into Driver classes which are responsible for preparing
driver-specific capabilities, calling handshake and using the bridge
it returns (OSS or W3C) to abstract dialect-specific communication.

For Firefox, it introduces Legacy and Marionette namespaces with its own
Driver classes and bridge extensions.

For Edge, it uses a hacky way to force W3C dialect even though
MicrosoftWebDriver responds with OSS-like body.

It also namespaces remote bridges/commands/etc. into OSS and W3C
modules respectively.
