+++ rb/lib/selenium/webdriver/edge.rb
@@ -19,8 +19,9 @@
 
 require 'net/http'
 
-require 'selenium/webdriver/edge/service'
 require 'selenium/webdriver/edge/bridge'
+require 'selenium/webdriver/edge/driver'
+require 'selenium/webdriver/edge/service'
 
 module Selenium
   module WebDriver
