+++ rb/lib/selenium/webdriver/edge/bridge.rb
@@ -20,42 +20,7 @@
 module Selenium
   module WebDriver
     module Edge
-      #
-      # @api private
-      #
-
-      class Bridge < Remote::W3CBridge
-        def initialize(opts = {})
-          opts[:desired_capabilities] ||= Remote::W3CCapabilities.edge
-
-          unless opts.key?(:url)
-            driver_path = opts.delete(:driver_path) || Edge.driver_path
-            port = opts.delete(:port) || Service::DEFAULT_PORT
-
-            opts[:driver_opts] ||= {}
-            if opts.key? :service_args
-              WebDriver.logger.warn <<-DEPRECATE.gsub(/\n +| {2,}/, ' ').freeze
-            [DEPRECATION] `:service_args` is deprecated. Pass switches using `driver_opts`
-              DEPRECATE
-              opts[:driver_opts][:args] = opts.delete(:service_args)
-            end
-
-            @service = Service.new(driver_path, port, opts.delete(:driver_opts))
-            @service.host = 'localhost' if @service.host == '127.0.0.1'
-            @service.start
-            opts[:url] = @service.uri
-          end
-
-          super(opts)
-        end
-
-        def browser
-          :edge
-        end
-
-        def driver_extensions
-          [DriverExtensions::TakesScreenshot]
-        end
+      module Bridge
 
         def commands(command)
           unsupported = %i[execute_script execute_async_script submit_element double_click
@@ -64,21 +29,10 @@ module Selenium
            get_window_size set_window_size get_window_position set_window_position
            maximize_window get_alert_text accept_alert dismiss_alert]
           if unsupported.include? command
-            Remote::Bridge::COMMANDS[command]
+            Remote::OSS::Bridge::COMMANDS[command]
           else
             super
           end
-
-        end
-
-        def capabilities
-          @capabilities ||= Remote::Capabilities.edge
-        end
-
-        def quit
-          super
-        ensure
-          @service.stop if @service
         end
 
         def send_keys_to_active_element(key)
@@ -116,49 +70,6 @@ module Selenium
           execute :maximize_window, window_handle: handle
         end
 
-        def create_session(desired_capabilities)
-          resp = raw_execute :new_session, {}, {desiredCapabilities: desired_capabilities}
-          @session_id = resp['sessionId']
-          return Remote::W3CCapabilities.json_create resp['value'] if @session_id
-
-          raise Error::WebDriverError, 'no sessionId in returned payload'
-        end
-
-        #
-        # executes a command on the remote server.
-        #
-        #
-        # Returns the 'value' of the returned payload
-        #
-
-        def execute(*args)
-          result = raw_execute(*args)
-          result.payload.key?('value') ? result['value'] : result
-        end
-
-        #
-        # executes a command on the remote server.
-        #
-        # @return [WebDriver::Remote::Response]
-        #
-
-        def raw_execute(command, opts = {}, command_hash = nil)
-          verb, path = commands(command) || raise(ArgumentError, "unknown command: #{command.inspect}")
-          path = path.dup
-
-          path[':session_id'] = @session_id if path.include?(':session_id')
-
-          begin
-            opts.each do |key, value|
-              path[key.inspect] = escaper.escape(value.to_s)
-            end
-          rescue IndexError
-            raise ArgumentError, "#{opts.inspect} invalid for #{command.inspect}"
-          end
-
-          WebDriver.logger.info("-> #{verb.to_s.upcase} #{path}")
-          http.call verb, path, command_hash
-        end
       end # Bridge
     end # Edge
   end # WebDriver
