+++ rb/lib/selenium/webdriver/edge/driver.rb
@@ -19,17 +19,21 @@
 
 module Selenium
   module WebDriver
-    module IE
+    module Edge
+
       #
+      # Driver implementation for Microsoft Edge.
       # @api private
       #
 
-      class Bridge < Remote::Bridge
+      class Driver < WebDriver::Driver
+        include DriverExtensions::TakesScreenshot
+
         def initialize(opts = {})
-          opts[:desired_capabilities] ||= Remote::Capabilities.internet_explorer
+          opts[:desired_capabilities] ||= Remote::W3C::Capabilities.edge
 
           unless opts.key?(:url)
-            driver_path = opts.delete(:driver_path) || IE.driver_path
+            driver_path = opts.delete(:driver_path) || Edge.driver_path
             port = opts.delete(:port) || Service::DEFAULT_PORT
 
             opts[:driver_opts] ||= {}
@@ -40,33 +44,28 @@ module Selenium
               opts[:driver_opts][:args] = opts.delete(:service_args)
             end
 
-            %i[log_level log_file implementation].each do |method|
-              next unless opts.key? method
-              WebDriver.logger.warn <<-DEPRECATE.gsub(/\n +| {2,}/, ' ').freeze
-            [DEPRECATION] `#{method}` is deprecated. Pass switches using `driver_opts`
-              DEPRECATE
-              opts[:driver_opts][method] = opts.delete(method)
-            end
-
             @service = Service.new(driver_path, port, opts.delete(:driver_opts))
+            @service.host = 'localhost' if @service.host == '127.0.0.1'
             @service.start
             opts[:url] = @service.uri
           end
 
-          if opts.delete(:introduce_flakiness_by_ignoring_security_domains)
-            opts[:desired_capabilities][:ignore_protected_mode_settings] = true
-          end
-          opts[:desired_capabilities][:native_events] = opts.delete(:native_events) != false
+          # Edge is mostly using W3C dialect, but a request to
+          # create session responds with OSS-like body,
+          # so we need to force W3C implementation.
+          desired_capabilities = opts.delete(:desired_capabilities)
+          bridge = Remote::Bridge.new(opts)
+          capabilities = bridge.create_session(desired_capabilities)
 
-          super(opts)
+          WebDriver.logger.info 'Forcing W3C dialect.'
+          @bridge = Remote::W3C::Bridge.new(capabilities, bridge.session_id, opts)
+          @bridge.extend Edge::Bridge
+
+          super(@bridge, listener: opts[:listener])
         end
 
         def browser
-          :internet_explorer
-        end
-
-        def driver_extensions
-          [DriverExtensions::TakesScreenshot]
+          :edge
         end
 
         def quit
@@ -74,7 +73,8 @@ module Selenium
         ensure
           @service.stop if @service
         end
-      end # Bridge
-    end # IE
+
+      end # Driver
+    end # Edge
   end # WebDriver
 end # Selenium
