+++ rb/lib/selenium/webdriver/firefox.rb
@@ -21,15 +21,17 @@ require 'timeout'
 require 'socket'
 require 'rexml/document'
 
+require 'selenium/webdriver/firefox/driver'
+
 require 'selenium/webdriver/firefox/util'
 require 'selenium/webdriver/firefox/extension'
 require 'selenium/webdriver/firefox/binary'
 require 'selenium/webdriver/firefox/profiles_ini'
 require 'selenium/webdriver/firefox/profile'
 require 'selenium/webdriver/firefox/launcher'
-require 'selenium/webdriver/firefox/bridge'
-require 'selenium/webdriver/firefox/w3c_bridge'
-require 'selenium/webdriver/firefox/binary'
+require 'selenium/webdriver/firefox/legacy/driver'
+
+require 'selenium/webdriver/firefox/marionette/driver'
 require 'selenium/webdriver/firefox/service'
 
 module Selenium
