+++ rb/lib/selenium/webdriver/ie/driver.rb
@@ -19,27 +19,38 @@
 
 module Selenium
   module WebDriver
-    module PhantomJS
+    module IE
+
       #
+      # Driver implementation for Internet Explorer supporting
+      # both OSS and W3C dialects of JSON wire protocol.
       # @api private
       #
 
-      class Bridge < Remote::Bridge
+      class Driver < WebDriver::Driver
+        include DriverExtensions::TakesScreenshot
+
         def initialize(opts = {})
-          opts[:desired_capabilities] ||= Remote::Capabilities.phantomjs
+          opts[:desired_capabilities] ||= Remote::Capabilities.internet_explorer
 
           unless opts.key?(:url)
-            driver_path = opts.delete(:driver_path) || PhantomJS.driver_path
+            driver_path = opts.delete(:driver_path) || IE.driver_path
             port = opts.delete(:port) || Service::DEFAULT_PORT
 
             opts[:driver_opts] ||= {}
-            if opts.key? :args
+            if opts.key? :service_args
               WebDriver.logger.warn <<-DEPRECATE.gsub(/\n +| {2,}/, ' ').freeze
-            [DEPRECATION] `:args` is deprecated. Pass switches using `driver_opts`
+            [DEPRECATION] `:service_args` is deprecated. Pass switches using `driver_opts`
               DEPRECATE
-              opts[:driver_opts][:args] = opts.delete(:args)
-            elsif opts[:desired_capabilities]['phantomjs.cli.args']
-              opts[:driver_opts][:args] = opts[:desired_capabilities]['phantomjs.cli.args']
+              opts[:driver_opts][:args] = opts.delete(:service_args)
+            end
+
+            %i[log_level log_file implementation].each do |method|
+              next unless opts.key? method
+              WebDriver.logger.warn <<-DEPRECATE.gsub(/\n +| {2,}/, ' ').freeze
+            [DEPRECATION] `#{method}` is deprecated. Pass switches using `driver_opts`
+              DEPRECATE
+              opts[:driver_opts][method] = opts.delete(method)
             end
 
             @service = Service.new(driver_path, port, opts.delete(:driver_opts))
@@ -47,19 +58,17 @@ module Selenium
             opts[:url] = @service.uri
           end
 
-          super(opts)
+          if opts.delete(:introduce_flakiness_by_ignoring_security_domains)
+            opts[:desired_capabilities][:ignore_protected_mode_settings] = true
+          end
+          opts[:desired_capabilities][:native_events] = opts.delete(:native_events) != false
+
+          @bridge = Remote::Bridge.handshake(opts)
+          super(@bridge, listener: opts[:listener])
         end
 
         def browser
-          :phantomjs
-        end
-
-        def driver_extensions
-          [DriverExtensions::TakesScreenshot]
-        end
-
-        def capabilities
-          @capabilities ||= Remote::Capabilities.phantomjs
+          :internet_explorer
         end
 
         def quit
@@ -67,7 +76,8 @@ module Selenium
         ensure
           @service.stop if @service
         end
-      end # Bridge
-    end # PhantomJS
+
+      end # Driver
+    end # IE
   end # WebDriver
 end # Selenium
