+++ rb/spec/unit/selenium/webdriver/edge/service_spec.rb
@@ -23,13 +23,12 @@ module Selenium
   module WebDriver
     module Edge
       describe Service do
-        let(:resp) { {'sessionId' => 'foo', 'value' => @default_capabilities} }
+        let(:resp) { {'sessionId' => 'foo', 'value' => Remote::Capabilities.edge.as_json} }
         let(:service) { double(Service, start: true, uri: 'http://example.com', host: 'localhost', binary_path: nil) }
-        let(:caps) { {} }
+        let(:caps) { Remote::Capabilities.edge }
         let(:http) { double(Remote::Http::Default, call: resp).as_null_object }
 
         before do
-          @default_capabilities = Remote::Capabilities.edge.as_json
           allow(Remote::Capabilities).to receive(:edge).and_return(caps)
           allow_any_instance_of(Service).to receive(:start)
           allow_any_instance_of(Service).to receive(:binary_path)
@@ -39,13 +38,13 @@ module Selenium
           expect(Service).not_to receive(:new)
           expect(http).to receive(:server_url=).with(URI.parse('http://example.com:4321'))
 
-          Bridge.new(http_client: http, url: 'http://example.com:4321')
+          Driver.new(http_client: http, url: 'http://example.com:4321')
         end
 
         it 'defaults to desired path and port' do
           expect(Service).to receive(:new).with(Edge.driver_path, Service::DEFAULT_PORT, {}).and_return(service)
 
-          Bridge.new(http_client: http)
+          Driver.new(http_client: http)
         end
 
         it 'accepts a driver path & port' do
@@ -53,7 +52,7 @@ module Selenium
           port = '1234'
           expect(Service).to receive(:new).with(path, '1234', {}).and_return(service)
 
-          Bridge.new(http_client: http, driver_path: path, port: port)
+          Driver.new(http_client: http, driver_path: path, port: port)
         end
 
         it 'accepts driver options' do
@@ -65,8 +64,8 @@ module Selenium
                   "–package=#{driver_opts[:package]}",
                   "-verbose"]
 
-          bridge = Bridge.new(http_client: http, driver_opts: driver_opts)
-          expect(bridge.instance_variable_get("@service").instance_variable_get("@extra_args")).to eq args
+          driver = Driver.new(http_client: http, driver_opts: driver_opts)
+          expect(driver.instance_variable_get("@service").instance_variable_get("@extra_args")).to eq args
         end
 
         it 'deprecates `service_args`' do
@@ -77,8 +76,8 @@ module Selenium
 
           message = /\[DEPRECATION\] `:service_args` is deprecated. Pass switches using `driver_opts`/
 
-          expect { @bridge = Bridge.new(http_client: http, service_args: args) }.to output(message).to_stdout_from_any_process
-          expect(@bridge.instance_variable_get("@service").instance_variable_get("@extra_args")).to eq args
+          expect { @driver = Driver.new(http_client: http, service_args: args) }.to output(message).to_stdout_from_any_process
+          expect(@driver.instance_variable_get("@service").instance_variable_get("@extra_args")).to eq args
         end
       end
     end # Edge
