+++ rb/spec/unit/selenium/webdriver/firefox/launcher_spec.rb
@@ -23,13 +23,12 @@ module Selenium
   module WebDriver
     module Firefox
       describe Launcher do
-        let(:resp) { {'sessionId' => 'foo', 'value' => @default_capabilities} }
-        let(:launcher) { double(Launcher, launch: true, url: 'http://example.com') }
-        let(:caps) { {} }
-        let(:http) { double(Remote::Http::Default, call: resp).as_null_object }
+        let(:resp) { {'sessionId' => 'foo', 'value' => Remote::Capabilities.firefox.as_json} }
+        let(:launcher) { instance_double(Launcher, launch: true, url: 'http://example.com') }
+        let(:caps) { Remote::Capabilities.firefox }
+        let(:http) { instance_double(Remote::Http::Default, call: resp).as_null_object }
 
         before do
-          @default_capabilities = Remote::Capabilities.firefox.as_json
           allow(Remote::Capabilities).to receive(:firefox).and_return(caps)
           allow_any_instance_of(Service).to receive(:start)
           allow_any_instance_of(Service).to receive(:binary_path)
@@ -39,20 +38,20 @@ module Selenium
           expect(Launcher).not_to receive(:new)
           expect(http).to receive(:server_url=).with(URI.parse('http://example.com:4321'))
 
-          Bridge.new(http_client: http, url: 'http://example.com:4321')
+          Driver.new(marionette: false, http_client: http, url: 'http://example.com:4321')
         end
 
         it 'defaults to desired port' do
           expect(Launcher).to receive(:new).with(anything, DEFAULT_PORT, nil).and_return(launcher)
 
-          Bridge.new(http_client: http)
+          Driver.new(marionette: false, http_client: http)
         end
 
         it 'accepts a driver port' do
           port = '1234'
           expect(Launcher).to receive(:new).with(anything, '1234', nil).and_return(launcher)
 
-          Bridge.new(http_client: http, port: port)
+          Driver.new(marionette: false, http_client: http, port: port)
         end
       end
     end # Firefox
