+++ rb/spec/unit/selenium/webdriver/phantomjs/driver_spec.rb
@@ -21,27 +21,28 @@ require File.expand_path('../../spec_helper', __FILE__)
 
 module Selenium
   module WebDriver
-    module Firefox
-      describe W3CBridge do
-        let(:launcher) { double(Launcher, launch: nil, quit: nil, url: 'http://localhost:4444/wd/hub') }
-        let(:resp) { { 'value' => {'sessionId' => 'foo', 'value' => @default_capabilities} } }
-        let(:http) { double(Remote::Http::Default, call: resp).as_null_object }
-        let(:caps) { Remote::Capabilities.chrome }
+    module PhantomJS
+      describe Driver do
+        let(:resp)    { {'sessionId' => 'foo', 'value' => Remote::Capabilities.phantomjs.as_json} }
+        let(:service) { instance_double(Service, start: true, uri: 'http://example.com') }
+        let(:http)    { instance_double(Remote::Http::Default, call: resp).as_null_object }
 
         before do
-          @default_capabilities = Remote::Capabilities.firefox.as_json
-          allow(Remote::Capabilities).to receive(:firefox).and_return(caps)
-          allow(Launcher).to receive(:new).and_return(launcher)
+          allow(Service).to receive(:binary_path).and_return('/foo')
+          allow(Service).to receive(:new).and_return(service)
         end
 
-        it 'accepts server URL' do
-          expect(Service).not_to receive(:new)
-          expect(http).to receive(:server_url=).with(URI.parse('http://example.com:4321'))
+        it 'takes desired capabilities' do
+          custom_caps = Remote::Capabilities.new(browser_name: 'foo')
 
-          W3CBridge.new(http_client: http, url: 'http://example.com:4321')
+          expect(http).to receive(:call) do |_verb, _post, payload|
+            expect(payload[:desiredCapabilities]).to eq(custom_caps)
+            resp
+          end
+
+          Driver.new(http_client: http, desired_capabilities: custom_caps)
         end
-
       end
-    end # Firefox
+    end # PhantomJS
   end # WebDriver
 end # Selenium
