+++ rb/spec/unit/selenium/webdriver/safari/service_spec.rb
@@ -23,13 +23,12 @@ module Selenium
   module WebDriver
     module Safari
       describe Service do
-        let(:resp) { {'sessionId' => 'foo', 'value' => @default_capabilities} }
-        let(:service) { double(Service, start: true, uri: 'http://example.com') }
-        let(:caps) { {} }
-        let(:http) { double(Remote::Http::Default, call: resp).as_null_object }
+        let(:resp) { {'sessionId' => 'foo', 'value' => Remote::Capabilities.safari.as_json} }
+        let(:service) { instance_double(Service, start: true, uri: 'http://example.com') }
+        let(:caps) { Remote::Capabilities.safari }
+        let(:http) { instance_double(Remote::Http::Default, call: resp).as_null_object }
 
         before do
-          @default_capabilities = Remote::Capabilities.safari.as_json
           allow(Remote::Capabilities).to receive(:safari).and_return(caps)
           allow_any_instance_of(Service).to receive(:start)
           allow_any_instance_of(Service).to receive(:binary_path)
@@ -39,13 +38,13 @@ module Selenium
           expect(Service).not_to receive(:new)
           expect(http).to receive(:server_url=).with(URI.parse('http://example.com:4321'))
 
-          Bridge.new(http_client: http, url: 'http://example.com:4321')
+          Driver.new(http_client: http, url: 'http://example.com:4321')
         end
 
         it 'defaults to desired path and port' do
           expect(Service).to receive(:new).with(Safari.driver_path, Service::DEFAULT_PORT, {}).and_return(service)
 
-          Bridge.new(http_client: http)
+          Driver.new(http_client: http)
         end
 
         it 'accepts a driver path & port' do
@@ -53,7 +52,7 @@ module Selenium
           port = '1234'
           expect(Service).to receive(:new).with(path, '1234', {}).and_return(service)
 
-          Bridge.new(http_client: http, driver_path: path, port: port)
+          Driver.new(http_client: http, driver_path: path, port: port)
         end
       end
     end # Safari
